+++ src/compas_rcf/ur/visualize_urscript.py
@@ -11,8 +11,8 @@ from compas.geometry import Transformation
 from compas_fab.robots.configuration import Configuration
 from compas_fab.robots.ur5 import Robot
 
-from compas_rcf.utils.compas_to_rhino import cgframe_to_rgplane
-from compas_rcf.utils.compas_to_rhino import matrix_to_rgtransform
+from compas_rcf.rhino import cgframe_to_rgplane
+from compas_rcf.rhino import matrix_to_rgtransform
 
 __all__ = ["visualize_urscript"]