+++ src/compas_rcf/utils/__init__.py
@@ -5,12 +5,6 @@ compas_rcf.utils
 
 .. currentmodule:: compas_rcf.utils
 
-RhinoCommon to COMPAS object conversions
-----------------------------------------
-Scripts for converting geometry objects between the two frameworks. Most of this
-has been integrated into COMPAS v0.15.0 but compas_fab needs earlier versions
-of COMPAS.
-
 JSON tools
 ----------
 Utility functions for loading and dumping
@@ -20,5 +14,10 @@ Assorted utilities
 ------------------
 When there's no other good place for a bit of code..
 
+.. toctree::
 
+   compas_rcf.utils.csv_reports
+   compas_rcf.utils.json_
+   compas_rcf.utils.ui
+   compas_rcf.utils.util_funcs
 """