+++ CHANGELOG.md
@@ -8,9 +8,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+*   MyPy configuration so Mypy can find imports when running from Vim + Ale. Libraries without typehints set to `ignore_missing_imports`.
+*   Class `compas_rcf.abb.helpers.RapidToolData` added to create Rapid tooldata from compas and Rhino geometries.
+*   Class property `weight` added to `compas_rcf.fabrication.ClayBullet`.
+
 ### Changed
 
+*   `compas_rcf.abb.helpers` split into `helpers` and `connectivity` make `helpers` importable from IronPython.
+*   Robot instruction in `compas_rcf.abb.programs.pick_bullet` to go to picking_frame now specifies `zone_pick` instead of `zone_travel` to make sure that the tool actuators activate at the correct time.
+*   Module `compas_rcf.utils.csv_reports` now adds more data from `ClayBullet` instances.
+*   `compas_rcf.install_rhino`, `compas_rcf.utils.compas_to_rhino` and `compas_rcf.rhino_to_compas` now moved to new package `compas_rcf.rhino` to more clearly show where an environment with Rhino is necessary. `install_rhino` is however still available from `compas_rcf.install_rhino`.
+
 ### Removed
+
 
 
 ## \[0.2.1\] \[2020-03-04\]